Added the VM

+++ src/debug.zig
@@ -7,13 +7,13 @@ const Value = @import("value.zig").Value;
 pub fn disassembleChunk(chunk: *Chunk, name: []const u8) void {
     std.debug.print("== {s} ==\n", .{name});
 
-    var offset: u32 = 0;
+    var offset: usize = 0;
     while (offset < chunk.bytecode.items.len) {
         offset = disassembleInstruction(chunk, offset);
     }
 }
 
-pub fn disassembleInstruction(chunk: *Chunk, offset: u32) u32 {
+pub fn disassembleInstruction(chunk: *Chunk, offset: usize) usize {
     std.debug.print("{d:0>4} ", .{offset});
 
     if (offset > 0 and chunk.lines.items[offset] == chunk.lines.items[offset - 1]) {
@@ -37,12 +37,12 @@ pub fn disassembleInstruction(chunk: *Chunk, offset: u32) u32 {
     }
 }
 
-inline fn simpleInstruction(name: []const u8, offset: u32) u32 {
+inline fn simpleInstruction(name: []const u8, offset: usize) usize {
     std.debug.print("{s}\n", .{name});
     return offset + 1;
 }
 
-inline fn constInstruction(name: []const u8, chunk: *Chunk, offset: u32) u32 {
+inline fn constInstruction(name: []const u8, chunk: *Chunk, offset: usize) usize {
     const constant = chunk.bytecode.items[offset + 1];
     std.debug.print(
         "{s:<16} {d:>4} '{any}'\n",
